feat: don't parse gamelist.xml if unchanged
Previously, if a gamelist.xml is large, there is a noticeable delay when opening a directory because the XML has to be read and parsed. If the gamelist.xml is not changed (we use file size as a fingerprint, maybe more metadata can be used as well), we can just load the entries from the database. With this change, we also have to make sure we update the database every time we parse the gamelist.xml.